restructure interpolators and update tests
